Let's begin talking computers with the kids!
This initial import contains a little history that did happen already,
and the awesome new "game" (not it's not really one yet) that we wrote
today with the kids.

The intent here is to grow the repository as we progress, both as a nice
history to share with the kids later (when they do discover `git`
maybe), and maybe to raise inspiration among daddy geeks.