Computer assisted homework checking.
This homework was about listing words that are inside boundaries as
defined by other words from the dictionnary. Of course, the exercice
themselves contains accentuated words that the kids have been told are
to be sorted exactly like unaccentuated ones, so the answers are wrong
and we could discuss in which way together.

Discussing how to sort accentuated characters being funnier than
actually trying to fix the code, I'll leave it at that.